% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta.R
\name{de.dt2dflux}
\alias{de.dt2dflux}
\title{DGE to Differential Fluxes}
\usage{
de.dt2dflux(
  model,
  de.res,
  topn = 100,
  padj.cutoff = 0.05,
  na2zero = TRUE,
  old = TRUE
)
}
\arguments{
\item{model}{metabolic map model, e.g. recon3d}

\item{de.res}{differential expression result as a data.table, with the
columns named id (gene symbols), log.fc (log fold-change), pval (raw p value)}

\item{topn}{at most top topn DE genes in either direction}

\item{padj.cutoff}{genes with BH-adjusted p value < padj.cutoff are kept as
the changed genes (i.e. with values -1/1), all other are set to 0 (unchanged)}

\item{na2zero}{if TRUE, NA's will be replaced by zeros in the final result}

\item{old}{old}
}
\value{
discretized dflux values
}
\description{
# map a differential expression analysis result given in de.res
to differential fluxes of reactions in a metabolic model
}
