% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta.R
\name{rmta}
\alias{rmta}
\title{the function for running rMTA (with MTA of the original MIQP version);}
\usage{
rmta(
  model,
  flux0,
  dflux,
  rxns = "all+ctrl",
  ko = NULL,
  nc = 1L,
  detail = TRUE,
  k = 100,
  mta.pars = list(),
  mip.pars = list(),
  qp.pars = list()
)
}
\arguments{
\item{model}{model}

\item{flux0}{flux}

\item{dflux}{dflux}

\item{rxns}{rxns}

\item{ko}{ko}

\item{nc}{nc}

\item{detail}{detail}

\item{k}{rMTA-specific parameter}

\item{mta.pars}{mta}

\item{mip.pars}{mip}

\item{qp.pars}{qp}
}
\value{
both the mta.model, and the summary data.table of MTA scores
for the rxns, in a list(mta.model, result.mta, result.moma, result.rmta)
}
\description{
the function for running rMTA (with MTA of the original MIQP version);
}
